feat: Modularize Neovim configuration
This commit refactors the kickstart.nvim configuration from a single
init.lua file into a modular structure. This enhances maintainability
and organization while preserving the beginner-friendly nature of the
project.

Key changes include:
- Core Neovim settings (options, keymaps, autocommands) are moved into
  `lua/kickstart/core/`.
- Plugin configurations are separated into individual files within
  `lua/kickstart/plugins/`.
- `init.lua` is updated to load these modules and now primarily serves as
  an orchestrator.
- README.md and code comments are updated to reflect the new structure,
  guiding you on how to navigate and customize your setup.
- The existing mechanism for custom plugins via `lua/custom/plugins/`
  is preserved and integrates seamlessly with the new structure.
- All new modules include top-level comments explaining their purpose.

This modularization follows best practices and addresses community
feedback, making it easier for you to understand, manage, and extend
your Neovim configuration using kickstart.nvim as a foundation.
