+++ lua/kickstart/plugins/autopairs.lua
@@ -1,5 +1,4 @@
--- autopairs
--- https://github.com/windwp/nvim-autopairs
+-- Configures nvim-autopairs for automatically managing pairs of characters.
 
 return {
   'windwp/nvim-autopairs',
