+++ lua/kickstart/plugins/debug.lua
@@ -1,10 +1,4 @@
--- debug.lua
---
--- Shows how to use the DAP plugin to debug your code.
---
--- Primarily focused on configuring the debugger for Go, but can
--- be extended to other languages as well. That's why it's called
--- kickstart.nvim and not kitchen-sink.nvim ;)
+-- Configures debugging capabilities using nvim-dap and related plugins.
 
 return {
   -- NOTE: Yes, you can install new plugins here!
