+++ lua/kickstart/plugins/guess-indent.lua
@@ -1,0 +1,4 @@
+-- Configures guess-indent.nvim for automatic indentation detection.
+return {
+  'NMAC427/guess-indent.nvim', -- Detect tabstop and shiftwidth automatically
+}
