+++ lua/kickstart/plugins/indent_line.lua
@@ -1,3 +1,4 @@
+-- Configures indent-blankline.nvim for displaying indent lines.
 return {
   { -- Add indentation guides even on blank lines
     'lukas-reineke/indent-blankline.nvim',
