+++ lua/kickstart/plugins/todo-comments.lua
@@ -1,0 +1,7 @@
+-- Configures todo-comments.nvim for highlighting TODO, FIXME, etc.
+return {
+  'folke/todo-comments.nvim',
+  event = 'VimEnter',
+  dependencies = { 'nvim-lua/plenary.nvim' },
+  opts = { signs = false },
+}
